tidy up GTF
